separated pool functions and cli
