+++ kalavai_client/assets/docker-compose-template.yaml
@@ -68,6 +68,8 @@ services:
     restart: unless-stopped
     command: >
       --command={{command}}
+      --node_name="{{node_name}}"
+      --node_ip="{{node_ip_address}}"
       {% if command == "server" %}
       --port_range="30000-32767"
       {% else %}
